Revise Begin to use primitive types in place of Unsigned variants
Use primitive types to reduce garbage related to boxing of Unsigned
types.  Future work need to validate mandatory fields on both encode and
decode of these types of objects.

+++ qpid-proton4j-engine/src/main/java/org/apache/qpid/proton4j/engine/impl/ProtonSession.java
@@ -64,7 +64,7 @@ public class ProtonSession extends ProtonEndpoint<Session> implements Session, P
     }
 
     public int getRemoteChannel() {
-        return remoteBegin != null ? remoteBegin.getRemoteChannel().intValue() : -1;
+        return remoteBegin != null ? remoteBegin.getRemoteChannel() : -1;
     }
 
     @Override
